wip: adapt creation for all activity types we expect to use
